LPS-32851 & LPS-33853 Extract common methods to super class (this methods will also be used in MemPolUserGroup and MemPolRoles tests)
